+++ atlas.com/pets/pet/entity.go
@@ -13,7 +13,7 @@ func Migration(db *gorm.DB) error {
 type Entity struct {
 	TenantId        uuid.UUID `gorm:"not null;"`
 	OwnerId         uint32    `gorm:"not null;"`
-	Id              uint32    `gorm:"primary_key;auto_increment"`
+	Id              uint64    `gorm:"primary_key;auto_increment"`
 	InventoryItemId uint32    `gorm:"not null"`
 	TemplateId      uint32    `gorm:"not null"`
 	Name            string    `gorm:"size:13;not null"`
