+++ atlas.com/pets/pet/provider.go
@@ -7,7 +7,7 @@ import (
 	"gorm.io/gorm"
 )
 
-func getById(tenantId uuid.UUID, id uint32) database.EntityProvider[Entity] {
+func getById(tenantId uuid.UUID, id uint64) database.EntityProvider[Entity] {
 	return func(db *gorm.DB) model.Provider[Entity] {
 		return database.Query[Entity](db, &Entity{TenantId: tenantId, Id: id})
 	}
